refactor: Favorite, Repository 책임 분리
레포지토리 서비스 레이어에서 같이 작성되었던 즐겨찾기 기능을 분리한다.
