+++ src/main/java/com/ky/docstory/controller/repository/RepositoryController.java
@@ -6,6 +6,7 @@ import com.ky.docstory.dto.repository.MyRepositoryResponse;
 import com.ky.docstory.dto.repository.RepositoryCreateRequest;
 import com.ky.docstory.dto.repository.RepositoryResponse;
 import com.ky.docstory.entity.User;
+import com.ky.docstory.service.favorite.FavoriteService;
 import com.ky.docstory.service.repository.RepositoryService;
 import jakarta.validation.Valid;
 import java.util.List;
@@ -21,6 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class RepositoryController implements RepositoryApi {
 
     private final RepositoryService repositoryService;
+    private final FavoriteService favoriteService;
 
     @Override
     public ResponseEntity<DocStoryResponseBody<RepositoryResponse>> createRepository(
@@ -38,20 +40,20 @@ public class RepositoryController implements RepositoryApi {
 
     @Override
     public ResponseEntity<DocStoryResponseBody<List<MyRepositoryResponse>>> getMyFavoriteRepositories(User currentUser) {
-        List<MyRepositoryResponse> myRepositoryResponses = repositoryService.getMyFavoriteRepositories(currentUser);
+        List<MyRepositoryResponse> myRepositoryResponses = favoriteService.getMyFavoriteRepositories(currentUser);
         return ResponseEntity.status(HttpStatus.OK).body(DocStoryResponseBody.success(myRepositoryResponses));
 
     }
 
     @Override
     public ResponseEntity<DocStoryResponseBody<MyRepositoryResponse>> addFavorite(User currentUser, UUID repositoryId) {
-        MyRepositoryResponse myRepositoryResponse = repositoryService.addFavorite(repositoryId, currentUser);
+        MyRepositoryResponse myRepositoryResponse = favoriteService.addFavorite(repositoryId, currentUser);
         return ResponseEntity.status(HttpStatus.OK).body(DocStoryResponseBody.success(myRepositoryResponse));
     }
 
     @Override
     public ResponseEntity<DocStoryResponseBody<Void>> removeFavorite(User currentUser, UUID repositoryId) {
-        repositoryService.removeFavorite(repositoryId, currentUser);
+        favoriteService.removeFavorite(repositoryId, currentUser);
         return ResponseEntity.status(HttpStatus.OK).body(DocStoryResponseBody.success(null));
     }
